fix list bug
