+++ homework/app/src/main/java/com/example/homework/item/FileItem.kt
@@ -4,9 +4,8 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import cn.nekocode.itempool.Item
 import com.example.homework.R
-import com.example.homework.data.DO.Filename
+import com.example.homework.data.DO.file.Filename
 import kotlinx.android.synthetic.main.item_file.view.*
-import kotlinx.android.synthetic.main.item_file_teachersfile.view.*
 
 /**
  * Created by Administrator on 2017/11/6 0006.
