+++ homework/app/src/main/java/com/example/homework/screen/file/myfile/MyFileActivity.kt
@@ -77,7 +77,6 @@ class MyFileActivity : BaseActivity(), Contract.View {
             intent.putExtra("cid",cid)
             intent.putExtra("name",name)
             startActivity(intent)
-            //this.overridePendingTransition(android.R.anim.fade_in,android.R.anim.fade_out);
         }
 
     }
