+++ homework/app/src/main/java/com/example/homework/screen/personal/main/PersonalActivity.kt
@@ -76,13 +76,11 @@ class PersonalActivity: BaseActivity(), Contract.View {
 
         tab_data.setOnClickListener {
             startActivity(Intent(this, FileActivity::class.java))
-            this.overridePendingTransition(android.R.anim.fade_in,android.R.anim.fade_out)
             finish()
         }
 
         tab_course.setOnClickListener {
             startActivity(Intent(this, CourseActivity::class.java))
-            this.overridePendingTransition(android.R.anim.fade_in,android.R.anim.fade_out)
             finish()
         }
     }
