+++ homework/app/src/main/res/layout/item_file_teachersfile.xml
@@ -21,14 +21,14 @@
         android:layout_weight="1">
 
         <TextView
-            android:id="@+id/txt_teachersfile_filename"
+            android:id="@+id/text_teachersfile_filename"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginLeft="9dp"
             android:layout_marginTop="9dp"/>
 
         <TextView
-            android:id="@+id/txt_teachersfile_time"
+            android:id="@+id/text_teachersfile_isdownload"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginLeft="13dp"/>
@@ -44,7 +44,7 @@
             android:id="@+id/img_teachersfile_isdownload"
             android:layout_width="20dp"
             android:layout_height="20dp"
-            android:src="@drawable/icon_teacher_file_download"/>
+            android:src="#ffffff"/>
     </LinearLayout>
 
 </LinearLayout>
