+++ homework/data/src/main/java/com/example/homework/data/DO/file/Check.kt
@@ -1,26 +1,26 @@
-package com.example.homework.data.DO
+package com.example.homework.data.DO.file
 
 import android.os.Parcel
 import android.os.Parcelable
+
 import com.google.gson.annotations.SerializedName
 import paperparcel.PaperParcel
 
 /**
- * Created by Administrator on 2017/11/17 0017.
+ * Created by Administrator on 2017/11/16 0016.
  */
 @PaperParcel
-data class SystemFile(
-        val name: String,
-        val path: String
+data class Check(
+        @SerializedName("status")val status:Int
 ) : Parcelable {
 
     companion object {
-        @JvmField val CREATOR = PaperParcelSystemFile.CREATOR
+        @JvmField val CREATOR = PaperParcelCheck.CREATOR
     }
 
     override fun describeContents() = 0
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
-        PaperParcelSystemFile.writeToParcel(this, dest, flags)
+        PaperParcelCheck.writeToParcel(this, dest, flags)
     }
 }
