+++ homework/data/src/main/java/com/example/homework/data/DO/file/ClassList.kt
@@ -1,26 +1,24 @@
-package com.example.homework.data.DO
+package com.example.homework.data.DO.file
 
 import android.os.Parcel
 import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
 import paperparcel.PaperParcel
 
 /**
  * Created by Administrator on 2017/11/16 0016.
  */
-
 @PaperParcel
-data class TeacherList(
-        val courseFile: ArrayList<TeachersFile>
+data class ClassList(
+        val classList: ArrayList<Filename>
 ) : Parcelable {
 
     companion object {
-        @JvmField val CREATOR = PaperParcelTeacherList.CREATOR
+        @JvmField val CREATOR = PaperParcelClassList.CREATOR
     }
 
     override fun describeContents() = 0
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
-        PaperParcelTeacherList.writeToParcel(this, dest, flags)
+        PaperParcelClassList.writeToParcel(this, dest, flags)
     }
 }
