+++ homework/data/src/main/java/com/example/homework/data/DO/file/TeachersFile.kt
@@ -1,4 +1,4 @@
-package com.example.homework.data.DO
+package com.example.homework.data.DO.file
 
 import android.os.Parcel
 import android.os.Parcelable
@@ -9,19 +9,19 @@ import paperparcel.PaperParcel
  * Created by Administrator on 2017/11/8 0008.
  */
 @PaperParcel
-data class MyFile(
-        @SerializedName("_id") override  var id: String,
-        @SerializedName("fname")var fname: String,
+data class TeachersFile(
+        @SerializedName("_id") var id: String,
+        @SerializedName("cfname")var filename: String,
         @SerializedName("url")var url: String
-) : WithId, Parcelable {
+) :Parcelable {
 
     companion object {
-        @JvmField val CREATOR = PaperParcelMyFile.CREATOR
+        @JvmField val CREATOR = PaperParcelTeachersFile.CREATOR
     }
 
     override fun describeContents() = 0
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
-        PaperParcelMyFile.writeToParcel(this, dest, flags)
+        PaperParcelTeachersFile.writeToParcel(this, dest, flags)
     }
 }
