+++ homework/data/src/main/java/com/example/homework/data/service/FileService.kt
@@ -1,7 +1,8 @@
 package com.example.homework.data.service
 
-import com.example.homework.data.DO.*
-import com.example.homework.data.DataLayer
+import com.example.homework.data.DO.file.Filename
+import com.example.homework.data.DO.file.MyFile
+import com.example.homework.data.DO.file.TeachersFile
 
 import com.example.homework.data.api.FileApi
 import io.reactivex.Observable
@@ -23,7 +24,6 @@ object FileService {
                     .subscribeOn(Schedulers.io())
                     .observeOn(Schedulers.io())
                     .map {
-                        println("==========${DataLayer.GSON!!.toJson(it)}")
                         it.data.classList
                     }
 
@@ -32,7 +32,6 @@ object FileService {
                     .subscribeOn(Schedulers.io())
                     .observeOn(Schedulers.io())
                     .map {
-                        println("==========${DataLayer.GSON!!.toJson(it)}")
                         it.data.courseFile
                     }
 
@@ -41,7 +40,6 @@ object FileService {
                     .subscribeOn(Schedulers.io())
                     .observeOn(Schedulers.io())
                     .map {
-                        println("==========${DataLayer.GSON!!.toJson(it)}")
                         it.data.selfFile
                     }
 
@@ -55,8 +53,6 @@ object FileService {
                .subscribeOn(Schedulers.io())
                .observeOn(Schedulers.io())
                .map {
-                   println("==========scuess")
-                   println("==========${DataLayer.GSON!!.toJson(it)}")
                    it.data.status
                }
     }
